requirements-frozen.txt
-e file:///Users/ravi/Documents/nullBytes_Apps/Ai_Agents/AI%20Ops
alembic==1.17.1
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.11.0
asgiref==3.10.0
asyncpg==0.30.0
attrs==25.4.0
authlib==1.6.5
bandit==1.8.6
bcrypt==4.3.0
beartype==0.22.5
bidict==0.23.1
billiard==4.2.2
black==25.9.0
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.1
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
configargparse==1.7.1
coverage==7.11.0
cryptography==46.0.3
cyclopts==4.2.1
deprecated==1.3.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22.3
durationpy==0.10
dydantic==0.0.8
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fakeredis==2.32.1
fastapi==0.120.4
fastmcp==2.13.0.2
flask==3.1.2
flask-cors==6.0.1
flask-login==0.6.3
gevent==25.9.1
geventhttpclient==2.3.5
gitdb==4.0.12
gitpython==3.1.45
google-auth==2.42.1
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.0
iniconfig==2.3.0
itsdangerous==2.2.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jiter==0.11.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.6.0
kombu==5.5.4
kubernetes==34.1.0
langchain==1.0.5
langchain-anthropic==1.0.1
langchain-core==1.0.4
langchain-mcp-adapters==0.1.12
langchain-openai==1.0.3
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-checkpoint-postgres==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langmem==0.0.30
langsmith==0.4.39
limits==5.6.0
locust==2.42.3
locust-cloud==1.29.0
loguru==0.7.3
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.20.0
mdurl==0.1.2
more-itertools==10.8.0
msgpack==1.1.2
mypy==1.18.2
mypy-extensions==1.1.0
narwhals==2.10.1
numpy==2.3.4
oauthlib==3.3.1
openai==2.6.1
openapi-pydantic==0.5.1
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-celery==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-instrumentation-httpx==0.59b0
opentelemetry-instrumentation-redis==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.4
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathable==0.4.4
pathspec==0.12.1
pathvalidate==3.3.1
pgvector==0.4.1
pillow==12.0.0
pip==25.3
platformdirs==4.5.0
playwright==1.55.0
plotly==6.3.1
pluggy==1.6.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
protobuf==6.33.0
psutil==7.1.3
psycopg==3.2.12
psycopg-pool==3.2.7
psycopg2-binary==2.9.11
py-key-value-aio==0.2.8
py-key-value-shared==0.2.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0
pydeck==0.9.1
pyee==13.0.0
pygments==2.19.2
pyjwt==2.10.1
pyperclip==1.11.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-base-url==2.1.0
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-httpx==0.35.0
pytest-json-report==1.5.0
pytest-metadata==3.1.1
pytest-playwright==0.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-engineio==4.12.3
python-jose==3.5.0
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.14.3
pytokens==0.2.0
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
redis==5.2.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
responses==0.25.8
respx==0.22.0
rich==14.2.0
rich-rst==1.3.2
rpds-py==0.28.0
rsa==4.9.1
ruff==0.14.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.44
sse-starlette==3.0.3
starlette==0.49.1
stevedore==5.5.0
streamlit==1.51.0
tenacity==9.1.2
text-unidecode==1.3
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
trustcall==0.0.39
types-cachetools==6.2.0.20251022
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250914
types-python-jose==3.5.0.20250531
types-pyyaml==6.0.12.20250915
types-zxcvbn==4.5.0.20250809
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.3.1
xxhash==3.6.0
zipp==3.23.0
zope-event==6.1
zope-interface==8.1.1
zstandard==0.25.0
zxcvbn==4.5.0
